Initial attempt at a bluetooth rfcomm port
Added a bluetooth port type that uses an RFCOMM socket.  Similar to the
existing TCP port, but with slightly less functionality for now (e.g.
cannot yet specify the bluetooth device to listen on).  The goal is to
be able to connect to a cheap serial/bluetooth module like the ones
people use with Arduinos, without having to set up an rfcomm device
node.  Only works with bluez on linux for now.

Also added a bluetooth port example that's a copy of the serial port
loopback device.  There's no software loopback to proper testing will
require a hardware loopback (e.g. bridging the tx and rx pins on a
bluetooth/serial module)